feat: moving fixture location for better context understanding
